+++ web/vtctld2/src/app/dashboard/dashboard.component.ts
@@ -14,8 +14,8 @@ import { PolymerElement } from '@vaadin/angular2-polymer';
 
 import { AddButtonComponent } from '../shared/add-button.component';
 import { DialogComponent } from '../shared/dialog/dialog.component';
-import { DialogContent } from '../shared/dialog/dialogContent';
-import { DialogSettings } from '../shared/dialog/dialogSettings';
+import { DialogContent } from '../shared/dialog/dialog-content';
+import { DialogSettings } from '../shared/dialog/dialog-settings';
 import { ForceFlag, KeyspaceNameFlag, RecursiveFlag, ShardingColumnNameFlag, ShardingColumnTypeFlag } from '../shared/flags/keyspace.flags';
 import { KeyspaceService } from '../api/keyspace.service';
 import { PrepareResponse } from '../shared/prepare-response';
