+++ web/vtctld2/src/app/dashboard/keyspace-extra.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { DialogContent } from '../shared/dialog/dialogContent';
+import { DialogContent } from '../shared/dialog/dialog-content';
 
 @Component({
   moduleId: module.id,
