+++ web/vtctld2/src/app/shared/dialog/dialog-content.ts
